+++ mlsm/test/test_fcn.py
@@ -38,3 +38,69 @@ def basic_sum_fcn_multiple(data, results):
     resultsReturn = {}
     resultsReturn['d'] = y
     return resultsReturn
+
+############################################################
+## Fcns for unit testing
+############################################################
+
+passes_intSingle_fieldset = {
+    'a': int
+}
+
+def passes_intSingle(data, results):
+    x = data['a'] + 1
+    resultsReturn = {}
+    resultsReturn['x'] = x
+    return resultsReturn
+
+passes_intMultiple_fieldset = {
+    'a': int,
+    'b': int
+}
+
+def passes_intMultiple(data, results):
+    x = data['a'] + data['b']
+    resultsReturn = {}
+    resultsReturn['x'] = x
+    return resultsReturn
+
+fails_intSingle_fieldset = {
+    'a': int
+}
+
+def fails_intSingle(data, results):
+    x = data['a'] + 'lorem ipsum'
+    resultsReturn = {}
+    resultsReturn['x'] = x
+    return resultsReturn
+
+passes_strSingle_fieldset = {
+    'a': str
+}
+
+def passes_strSingle(data, results):
+    x = data['a'] + ' test'
+    resultsReturn = {}
+    resultsReturn['x'] = x
+    return resultsReturn
+
+passes_strMultiple_fieldset = {
+    'a': str,
+    'b': str
+}
+
+def passes_strMultiple(data, results):
+    x = data['a'] + ' ' + data['b']
+    resultsReturn = {}
+    resultsReturn['x'] = x
+    return resultsReturn
+
+fails_strSingle_fieldset = {
+    'a': str
+}
+
+def fails_strSingle(data, results):
+    x = data['a'] + 1
+    resultsReturn = {}
+    resultsReturn['x'] = x
+    return resultsReturn