+++ mlsm/test/test_mlsm.py
@@ -280,3 +280,37 @@ def test_RunSummaryWrongVersion():
     summodellist = [{'name': 'test1', 'version': '0.0.1'}]
     testsummodel = mlsm.SummaryModel(name='testSummary', models = summodellist, fields=test_fcn.basic_fcn_add_fieldset, version=version, fcn = test_fcn.basic_sum_fcn_multiple)
     testResults = mlsm.RunModelsAll(models = [testmodel1], summaryModels=[testsummodel], records = testData)
+
+###############################################################################
+## Model unit testing to make sure things go OK
+###############################################################################
+
+def test_RunModelTest_Passes_IntSingle():
+    testmodel = mlsm.Model(name='test', fields=test_fcn.passes_intSingle_fieldset, version=version, fcn=test_fcn.passes_intSingle)
+    result = testmodel.test()
+    assert result
+
+def test_RunModelTest_Passes_IntMultiple():
+    testmodel = mlsm.Model(name='test', fields=test_fcn.passes_intMultiple_fieldset, version=version, fcn=test_fcn.passes_intMultiple)
+    result = testmodel.test()
+    assert result
+
+def test_RunModelTest_Fails_IntSingle():
+    testmodel = mlsm.Model(name='test', fields=test_fcn.fails_intSingle_fieldset, version=version, fcn=test_fcn.fails_intSingle)
+    result = testmodel.test()
+    assert not result
+
+def test_RunModelTest_Passes_StrSingle():
+    testmodel = mlsm.Model(name='test', fields=test_fcn.passes_strSingle_fieldset, version=version, fcn=test_fcn.passes_strSingle)
+    result = testmodel.test()
+    assert result
+
+def test_RunModelTest_Passes_StrMultiple():
+    testmodel = mlsm.Model(name='test', fields=test_fcn.passes_strMultiple_fieldset, version=version, fcn=test_fcn.passes_strMultiple)
+    result = testmodel.test()
+    assert result
+
+def test_RunModelTest_Fails_StrSingle():
+    testmodel = mlsm.Model(name='test', fields=test_fcn.fails_strSingle_fieldset, version=version, fcn=test_fcn.fails_strSingle)
+    result = testmodel.test()
+    assert not result